feat: Add Docker Compose file for MongoDB
Add a new Docker Compose file to the explorer_api directory to set up a MongoDB container. The file includes the necessary environment variables for initializing the MongoDB database. This will allow for easier deployment and testing of the application.

Also, update the package.json file to include new dependencies: axios and dotenv. These packages will be used for making HTTP requests and managing environment variables, respectively.

Finally, update the index.js file to include new routes for retrieving blocks and block details. The /blocks route will be used to retrieve a list of all blocks, while the /block/:id route will retrieve details for a specific block using an HTTP POST request to the specified node URL.

This commit addresses the need for a MongoDB container, adds necessary dependencies, and expands the functionality of the API.